Add logic for memory chip.
Memory chip of a main RAM16K memory, a screen and a keyboard.

Based upon address[14] and address[13] its decided if we'll use RAM,
Screen or Keyboard by a dmux.

Since RAM16K uses 14 address lines ( 0 to 13) . So number of addresses
are 2^14 = 16384.
Screen uses only 13 address lines (0 to 12)
And keyboard doesn't use any address line

The outputs of these memories are mux'd similarly as the dmux, mux
sandwitch...