Add logic foor implementation of CPU.
Use the tests as CPU-extenal for testing these.

This must have been the most difficult to implement chip so far. However
there is a core abstraction of CPU at work which tok me some time to
realize.

A CPU interface is clear:
It takes an instruction from
somewhere (typically program memory), it
takes an input word from somewhere (typically data memory) and takes the
reset input.

Then next it gives what should be the next instruction address, if
something should be written to memory then what should it be and where
should it be written on the memory.

Implementation:
Clearly CPU encapsulates D and A register and I think they are valid
places to start with.

D register is loaded if its a C instruction and d2 bit is asserted.
A is loaded if its a A instructio or d1 bit is asserted.

The in word of D register is always output of ALU while in word of A
register is either the output of ALU (depending if its was a valid
destination through C instruction) or the instruction itself if it was
an A instruction.

Now the inputs to the ALU (x and y) x is same as D. while y is M/A and
has to be mux'd and comes from the 'a' bit of C insturction.

c1..c6 bits of C instruction forms control bits of ALU.

Jump bits are used as select lines in a mux logic'd with ALU output
control bits to see if there should be a jump, and ofcurse if it was a C
instruction. If so program counter jumps to value in A register else inc
is always true while reset is controlled...

And finally we see if we need to write to the memory (C-instruction and
d3)

Wow.